+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This release changes some API methods.
 - Feat: add more inline help text's #1072
 - Feat: improve validation of form fields #1083
 - Feat: add spinners to buttons
+- Feat: display the number of channels #1085
 - Upd: some code enhancements #1045
 - Upd: Bootstrap 5.3.1
 - Upd: rework settings